자잘한거

+++ src/main/java/org/zerock/controller/ReplyController.java
@@ -1,6 +1,5 @@
 package org.zerock.controller;
 
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
 
 import java.security.Principal;
 import java.util.List;
@@ -8,15 +7,11 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.RestController;
 import org.zerock.domain.Criteria;
 import org.zerock.domain.ReplyVo;
 import org.zerock.service.ReplyService;
@@ -35,11 +30,17 @@ public class ReplyController {
 	@PostMapping(value = "/insert")
 	public String insert(ReplyVo vo , Principal principal) {
 		log.info("insert Vo = " + vo);
+		
 		vo.setMem_id(principal.getName());
+		
 		log.info("insert Vo = " + vo);
+		
 		int result = service.insert(vo);
+		
 		log.info("°á°ú  = " + result);
+		
 //		result == 1 ? "success" : "error";
+		
 		return "redirect:/board/get?board_no="+vo.getBoard_no();
 	}
 
@@ -53,7 +54,7 @@ public class ReplyController {
 		log.info(list);
 		return list;
 	}
-
+	
 	@GetMapping(value = "/{board_no}/{reply_no}", produces = { MediaType.APPLICATION_JSON_UTF8_VALUE })
 	@ResponseBody
 	public ReplyVo get(@PathVariable("reply_no") int reply_no) {
@@ -63,15 +64,20 @@ public class ReplyController {
 	}
 	
 	@PostMapping(value = "/{board_no}/{reply_no}")
-	public String delete(@PathVariable("board_no") int board_no , @PathVariable("reply_no") int reply_no) {
+	public String delete(@PathVariable("board_no") Long board_no , @PathVariable("reply_no") int reply_no) {
 		log.info("delete" + reply_no);
+		
 		int result = service.delete(reply_no);
+		
 		return "redirect:/board/get?board_no="+board_no;
 	}
+	
 	@PostMapping(value = "/{board_no}/update")
-	public String update(@PathVariable("board_no") int board_no , ReplyVo vo) {
+	public String update(@PathVariable("board_no") Long board_no , ReplyVo vo) {
 		log.info("update" + vo);
+		
 		int result = service.update(vo);
+		
 		return "redirect:/board/get?board_no="+board_no;
 	}
 }
